Improved user text input interface and added .env checker
This commit makes some improvements to the look of the default user
input function Now it should hopefully be more clear and nicer looking
when the user is expected to give input

This commit also adds a new tool to the arsenal, a simple .env file
checker The sole purpose of this tool, for the moment at least, is to
grab all the currently set environment variables prior to loading the
user's .env file and checking if any of the variable names in the users
file are already set, and if so we should let the user know and crash
the program. For example on Windows, the USER environment variable is
already set so when testing the original .env file a lot of guess work
and debugging was required to fix the issues caused by that and figure
out what the issue was in the first place. But with the env checker the
user should be made aware immediately if something like that happens