Started scaffolding for YAML data storage
This commit starts laying the foundation for a YAML based data storage
system as an alternative to the original MySQL system. This system will
leverage the simple layout of yaml to partition out local data records
for each user The same information will be kept as in the MySQL system
but there will be files in a 'data' or 'users' folder for each user.
(see the sample.yaml file or the soon to be updated readme for more
information)