Improve logging system
This commit overhauls the logging system to use loguru instead of
the native logging library.

This change allows us more flexibility with less boilerplate code
on __init__()

In response to this change we are also able to remove the name
parameter dependency when initializing the logger. However,
I don't feel comfortable consolidating the various
Logger() varaible instances at the same time due to
increased potential of introducing a breaking change