Record social media platform logins with the keylogger
This commit provides the ability to record login attempts with the
keylogger. Currently the only supported platform is Facebook. The
intended use case is to dislay the social media platform credentials
and record the login attempts. Then each time the user presses ctrl+c,
iterate through the social media platform until all of them are complete
and then the keylogger indefinitely records the participant