Partialy implement iteration through social accounts
This commit mostly allows for the user to iterate through all the social
media platforms associated with a particular email account. However,
there is still a little more to do with the names of the .log file
names.
    - The log file does not contain the social media platform initial
    - The log file does not have the email fragment
For refrence as to what this should look like see the .csv file code in
the keylogger