Remove extraneous file generation by the keylogger
Previously the keylogger would generate extra .log files in an effort to
compartmentalize the tracking for each platform and email address.
Often times this was just extra cruft and did not always work as
expected
This commit removes this extraneous file generation and instead keeps
around a global .log file regardless of the platform or email address.
To differentiate between the different platforms, the keylogger
generates .csv files for each platform and also includes a fragment of
the email address used in the file. There is still a few things to
cleanup such as that the individual csv files do not have their
headings, and there is still a global .csv file being generated.
However, these are relatively harmless so I am okay with the current
state