Add a function to record keystrokes tailored for windows
We cannot use the native KeyboardInterrupt strategy on the native windows terminal for some reason.
This commit adds a new function for reccording which only windows platforms will use.
Instead of using the normal strategy, the user chooses the platform they are typing for first, and then once they are done, they close the window and repeat the process manually for the other platforms