Add a headless web server for kmlog to remotely start the keylogger
This tool was added so that the keylogger can be started remotely
on researcher controlled pages on the local machine through a web
request. The web server starts at port 5000 and has two endpoints:

1. start-server: As part of the JSON POST request to this endpoint,
the payload is expected to have a body propery with a `platform_id`
field. The acceptable options for this field are (and what platform they represent):
- 0: Facebook
- 1: Instagram
- 2: Twitter

2. end-server: This endpoint expects a standard POST request payload and
does not require a body property. When this endpoint is hit, it will kill
the server and thus, the keylogger if it is running